+++ CHANGELOG.rst
@@ -55,7 +55,7 @@ Added
 -  Add functionality to add alias commands just like a normal command but a string as ``command`` value, which points to
    a ``command_name``. Additionally ``title``, ``description``, ``hidden`` and ``group`` can be set.
 -  Autogenerate ResT for all commands with ``/commands rst``, but be aware that double whitespace are not printed. You
-   get ``\ \ `` instead, which can be replaced.
+   get "\\ \\ " instead, which can be replaced.
 
 Changes
 ~~~~~~~